GH-3397: Implemented the HTTP-based authentication for Git in Electron.
If a git operation (fetch, pull, merge, ...) requires authentication,
the Theia backend will ask the frontend for the username and password.

User credentials are not stored, re-used.

In electron, if a Git operation fails due to authentication, we
report it back to the user. This PR does not change the behavior in the
browser.

Closes: #3397

Signed-off-by: Akos Kitta <kittaakos@typefox.io>